remove utils to single package
