+++ packages/cli/package.json
@@ -37,7 +37,8 @@
     "request": "^2.88.2",
     "semver": "^6.3.0",
     "ts-node": "^8.9.1",
-    "yaml": "^1.7.2"
+    "yaml": "^1.7.2",
+    "@pipcook/pipcook-utils": "^0.5.0"
   },
   "gitHead": "1e21ad0799e221dcf4b7aa3a3b2187d5b8d8b865",
   "devDependencies": {
