+++ packages/core/package.json
@@ -16,17 +16,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@types/cli-progress": "^3.4.2",
-    "@types/fs-extra": "^8.0.1",
-    "@types/uuid": "^7.0.2",
-    "cli-progress": "^3.6.0",
-    "extract-zip": "^1.6.7",
-    "fs-extra": "^8.1.0",
-    "request": "^2.88.2",
-    "systeminformation": "^4.16.0",
-    "targz": "^1.0.1",
-    "uuid": "^3.4.0",
-    "xml2js": "^0.4.22"
+    "rxjs": "^6.5.3"
   },
   "gitHead": "1e21ad0799e221dcf4b7aa3a3b2187d5b8d8b865",
   "devDependencies": {
