+++ packages/daemon/package.json
@@ -12,6 +12,7 @@
     "uuid": "^8.0.0",
     "uuid-validate": "^0.0.3",
     "@pipcook/pipcook-core": "^0.5.16",
+    "@pipcook/pipcook-utils": "^0.5.0",
     "glob": "^7.1.6",
     "glob-promise": "^3.4.0",
     "fs-extra": "^8.1.0",
