+++ packages/plugins/data-access/pascalvoc-data-access/package.json
@@ -18,7 +18,8 @@
   "dependencies": {
     "@pipcook/pipcook-core": "^0.5.16",
     "glob": "^7.1.6",
-    "glob-promise": "^3.4.0"
+    "glob-promise": "^3.4.0",
+    "@pipcook/pipcook-utils": "^0.5.0"
   },
   "gitHead": "1e21ad0799e221dcf4b7aa3a3b2187d5b8d8b865",
   "devDependencies": {
