+++ packages/plugins/data-access/pascalvoc-data-access/src/index.ts
@@ -3,7 +3,8 @@
  * the data is conform to expectation.
  */
 
-import { ArgsType, parseAnnotation, DataAccessType, VocDataset, ImageDataLoader, ImageLabel } from '@pipcook/pipcook-core';
+import { ArgsType, DataAccessType, VocDataset, ImageDataLoader, ImageLabel } from '@pipcook/pipcook-core';
+import { parseAnnotation } from '@pipcook/pipcook-utils';
 import glob from 'glob-promise';
 import * as path from 'path';
 import * as fs from 'fs';
