+++ packages/plugins/data-collect/csv-data-collect/package.json
@@ -23,7 +23,8 @@
     "csv-stringify": "^5.3.6",
     "glob": "^7.1.6",
     "glob-promise": "^3.4.0",
-    "uuid": "^7.0.2"
+    "uuid": "^7.0.2",
+    "@pipcook/pipcook-utils": "^0.5.0"
   },
   "gitHead": "1e21ad0799e221dcf4b7aa3a3b2187d5b8d8b865",
   "devDependencies": {
