+++ packages/plugins/data-collect/csv-data-collect/src/index.ts
@@ -1,7 +1,8 @@
 /**
  * @file For plugin to collect test classification data
  */
-import { DataCollectType, ArgsType, download, unZipData } from '@pipcook/pipcook-core';
+import { DataCollectType, ArgsType } from '@pipcook/pipcook-core';
+import { unZipData, download } from '@pipcook/pipcook-utils';
 import * as path from 'path';
 import * as assert from 'assert';
 import * as fs from 'fs-extra';
