+++ packages/plugins/data-collect/image-classification-data-collect/package.json
@@ -21,7 +21,8 @@
     "fs-extra": "^8.1.0",
     "glob": "^7.1.6",
     "glob-promise": "^3.4.0",
-    "uuid": "^7.0.2"
+    "uuid": "^7.0.2",
+    "@pipcook/pipcook-utils": "^0.5.0"
   },
   "gitHead": "1e21ad0799e221dcf4b7aa3a3b2187d5b8d8b865",
   "devDependencies": {
