+++ packages/plugins/data-collect/mnist-data-collect/package.json
@@ -21,7 +21,8 @@
     "@types/cli-progress": "^3.4.2",
     "cli-progress": "^3.6.0",
     "jimp": "^0.10.0",
-    "mnist": "^1.1.0"
+    "mnist": "^1.1.0",
+    "@pipcook/pipcook-utils": "^0.5.0"
   },
   "gitHead": "1e21ad0799e221dcf4b7aa3a3b2187d5b8d8b865",
   "devDependencies": {
