+++ packages/plugins/data-collect/object-detection-coco-data-collect/src/index.ts
@@ -3,7 +3,8 @@
  * the data is conform to expectation.
  */
 
-import { ArgsType, unZipData, download, createAnnotationFromJson, DataCollectType } from '@pipcook/pipcook-core';
+import { ArgsType, DataCollectType } from '@pipcook/pipcook-core';
+import { unZipData, download, createAnnotationFromJson } from '@pipcook/pipcook-utils';
 import glob from 'glob-promise';
 import * as path from 'path';
 import * as assert from 'assert';
