+++ packages/plugins/data-collect/object-detection-pascolvoc-data-collect/src/index.ts
@@ -1,4 +1,5 @@
-import { ArgsType, unZipData, download, DataCollectType } from '@pipcook/pipcook-core';
+import { ArgsType, DataCollectType } from '@pipcook/pipcook-core';
+import { unZipData, download } from '@pipcook/pipcook-utils';
 import * as path from 'path';
 import * as assert from 'assert';
 import * as fs from 'fs-extra';
