+++ packages/utils/.npmignore
@@ -1,0 +1,3 @@
+node_modules
+.vscode/
+tsconfig.tsbuildinfo
