+++ packages/utils/package.json
@@ -1,0 +1,44 @@
+{
+  "name": "@pipcook/pipcook-utils",
+  "version": "0.5.0",
+  "description": "this is some public utils to be used by plugins and other parts of pipcook",
+  "main": "dist/index",
+  "types": "dist/index",
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "test": "ts-node ../../run_tests.ts",
+    "build": "npm run clean && npm run compile",
+    "clean": "rm -rf ./dist && rm -rf tsconfig.tsbuildinfo",
+    "compile": "tsc -b tsconfig.json"
+  },
+  "author": "",
+  "license": "Apache-2.0",
+  "dependencies": {
+    "@types/cli-progress": "^3.4.2",
+    "@types/extract-zip": "^1.6.2",
+    "@types/fs-extra": "^8.0.1",
+    "@types/request": "^2.48.5",
+    "@types/targz": "^1.0.0",
+    "@types/uuid": "^7.0.2",
+    "@types/xml2js": "^0.4.5",
+    "cli-progress": "^3.6.0",
+    "extract-zip": "^1.6.7",
+    "fs-extra": "^8.1.0",
+    "request": "^2.88.2",
+    "targz": "^1.0.1",
+    "uuid": "^3.4.0",
+    "xml2js": "^0.4.22"
+  },
+  "gitHead": "1e21ad0799e221dcf4b7aa3a3b2187d5b8d8b865",
+  "devDependencies": {
+    "@types/jasmine": "^3.5.7",
+    "jasmine": "^3.5.0",
+    "nyc": "14.1.1",
+    "ts-node": "^8.6.2"
+  },
+  "publishConfig": {
+    "access": "public"
+  }
+}
